add invalid characters to the no_windows_filenames hook
Summary:
If a filename passes this hook it should be checkoutable on windows.

Unfortunately it's not the case now - we're missing the check for invalid
characters. Although this list might be redundant and some of chars might be
already banned by other hook let's make this hook a place for all
windows-specific filename checks.

Reviewed By: farnz

Differential Revision: D26979418

fbshipit-source-id: f1fa685b9e7e5413d8030d18bc083458b6a148e1