mononoke/fileblob: create tempfiles in blobstore location
Summary:
Right now, fileblob crashes if those two things are on different devices
(because you cannot atomically rename across devices), so you need to make sure
your TMPDIR is on the same volume as your fileblob.

This is kinda annoying and kinda unnecessary. Let's just fix default to putting
the temporary files into the same location as the blobstore.

Note: while the temp files will be in the same directory as the rest of our
blobs, they don't have the `blob-` prefix (their prefix will be `.tmp`), so
they cannot be read by accident as if they were blobs.

Reviewed By: farnz

Differential Revision: D27046889

fbshipit-source-id: c2b47cd6927eef34ac19325f87f446a6f6532eaf