scs_server: add context for service errors
Summary:
Add a `.context` method for `ServiceError`, which allows the addition of
context information in errors.

Since these are wrapped Thrift errors, we can't use the usual error-chain
mechanism of `std::error::Error`.  Instead, we just prepend the message that
the Thrift client will see with the context.

Add an extenstion to `Result` for results that contain an error that can be
converted into a `ServiceError` to allow the addition of context when
processing a chain of `Result`s.

Reviewed By: StanislavGlebik

Differential Revision: D27079921

fbshipit-source-id: a1200f44346530c91bd559f4be0ca2b04f7d4480