scs_server: log when requests are cancelled
Summary:
When requests are cancelled, their futures are dropped without completion.
Currently this causes no logs or statistics to be logged, as normally this
would happen after the request implementation completes.

Add logging for cancelled requests.  Include the gathered statistics so far,
so that we know how much time was spent on the cancelled request.

Reviewed By: StanislavGlebik

Differential Revision: D27084866

fbshipit-source-id: d4c5c276d496478f0c7caa700627b92d8f9e80a2