segmented_changelog: inline update::build
Summary:
`update::build` is only used by the Seeder. The steps in this function are not
isolated enough from the seeder to have a separate function. The seeder has the
role of builing it's own type of StartState. It is also the only process that
deals with the IdMapVersionStore. The seeder is particular enough that it makes
sense to inline it's build order.

Reviewed By: quark-zju

Differential Revision: D27099265

fbshipit-source-id: f86b8d7d4637a5f2582e70fc58b60c2041b93548