mononoke/bonsai_hg_mapping: use rendez vous queries
Summary:
Like it says in the title. See the previous diff for context. This is a little
more complex than I'd like, and IMO largely had to do with the fact that we
have one method for mapping Bonsais AND hg, and that's a bit messy. I think
this is worth refactoring, but in the interest of keeping the diffs small, this
right now just ports the current API to use rendezvous under the hood.

Reviewed By: StanislavGlebik

Differential Revision: D27010315

fbshipit-source-id: 9bef893dd37db244b0b91f5de4fa79d0fdc47228