mononoke: use unlink() in PackBlob::put_packed
Summary:
The pack key needs to be removed using unlink() after target keys are pointed to it so that old unused packs can be GCed. e.g.

packer run 1: keys A and B are packed to P1
packer run 2: keys A and C are packed to P2
packer run 3: keys B and D are packed to P3

If we don't unlink P1 during run1 then GC can't collect the old pack after run 3 as the key will keep it live

Reviewed By: farnz

Differential Revision: D27188108

fbshipit-source-id: bbe22a011abbf85370a8c548413401ffb54b16b6