bookmarks: remove dangerous overrides
Summary:
Remove dangerous overrides used in the tests in the bookmarks crate.

The test was using dangerous overrides to change the blobstore, which is now
supported directly by the test repo factory.

It was also using dangerous overrides to override the bookmark update log to
reset its database so it looks like the log is empty.  Instead, clear out the
bookmark update log database as part of the test.

Reviewed By: StanislavGlebik

Differential Revision: D27169426

fbshipit-source-id: 64e1e89e31f62dcb585741ea728ebbe45f60fd38