mononoke: record which bonsai commits were pushed in hg sync bundle
Summary:
In the next diff I'd like to add support for syncing globalrev to our darkstorm
repos. Doing it the same way we do it for hgsql isn't going to work, because
darkstorm repos stores globalrevs the same way mononoke does it (i.e. per
commit entry in mysql) and not the way hgsql does (i.e. one row per repo).

In this diff I do a small refactoring that remembers which bonsai commits were pushed
in a bundle, so that in the next diff we can start writing them to darkstorm
db.

Reviewed By: krallin

Differential Revision: D27268778

fbshipit-source-id: bbb39de233719c8435d11d00980f6eaf5b755ba6